requirements.txt
requests==2.32.3
beautifulsoup4==4.13.3
lxml==5.3.2
fastapi==0.115.12
uvicorn==0.34.1
requests==2.32.3
python-dotenv==1.1.0
streamlit==1.44.1